% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testclass.R
\name{createTextData}
\alias{createTextData}
\title{Define constructor function for creating text data objects (S3) for testing purposes}
\usage{
createTextData(text)
}
\arguments{
\item{text}{User inputted text}
}
\value{
list of \code{text}
}
\description{
Define constructor function for creating text data objects (S3) for testing purposes
}
\examples{
\dontrun{
createTextData(processed_text)
}
}
