% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyClean.R
\name{easyClean}
\alias{easyClean}
\title{easyClean}
\usage{
easyClean(text)
}
\arguments{
\item{text}{User's chosen file or URL link of type htm or html that has been converted into a variable.}
}
\value{
A preprocessed matrix of \code{text}
}
\description{
easyClean cleans the text data. It makes the text lowercase, removes numbers, removes punctuation, removes numbers, and removes common English stopwords. It will then create a term document matrix. Its function is called and performed for easyRead.
}
\examples{
\dontrun{
# This function is executed in easyClean
}
}
