% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyFrequency.R
\name{easyFrequency}
\alias{easyFrequency}
\title{easyFrequency}
\usage{
easyFrequency(word_matrix, pos_words, neg_words)
}
\arguments{
\item{word_matrix}{A preprocessed term document matrix that has been converted into a matrix.}

\item{pos_words}{Positive word list from the Bing Dictionary}

\item{neg_words}{Negative word list from the Bing Dictionary}
}
\value{
A list containing word frequency, positive word frequency, negative word frequency, all positive words, count of positive words, all negative words, count of negative words
}
\description{
Gets the total frequency of words overall but also determines the frequency of positive and negative words found within the text.
}
\examples{
\dontrun{
easyFrequency(cleanText, pos, neg)
}
}
