% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyRead.R
\name{easyRead}
\alias{easyRead}
\title{easyRead}
\usage{
easyRead(file_path)
}
\arguments{
\item{file_path}{User's file within their chosen directory or a URL link of type htm or html}
}
\value{
A preprocessed matrix of \code{file_path}
}
\description{
Reads in the file the user selected and preprocesses it to determine what type of file or URL it is. The function will prepare the text to then be cleaned through easyClean which is called within easyRead.
}
\examples{
\dontrun{
easyRead("http://www.historyplace.com/speeches/anthony.htm")
# Output: cleaned text
}
}
