% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyWordCloud.R
\name{easyWordCloud}
\alias{easyWordCloud}
\title{easyWordCloud}
\usage{
easyWordCloud(freq_data, color_scheme = "green", top_n = 50)
}
\arguments{
\item{freq_data}{A dataframe.}

\item{color_scheme}{A color.}

\item{top_n}{A number for how many words should appear.}
}
\value{
A wordcloud of \code{freq_data} with optional color selection \code{color_scheme} and number of words shown \code{top_n}
}
\description{
Visualize your findings via word clouds courtesy of the wordcloud2 package. Enter current iteration of text data after it has been run through the easyFrequency function. Users must convert the list output of easyFrequency to dataframe for the function to work properly.
}
\examples{
\dontrun{
easyWordCloud(positive.df, "green", 7)
}
}
